Backport latest hmpps-template-kotlin updates
Bringing over recent changes introduced in the `hmpps-template-kotlin` which this project used as its base, which includes bump to certain dependencies.

The linter version was updated and as a consequence there are new style rules which had to be applied in many classes.
More info here: https://mojdt.slack.com/archives/C69NWE339/p1738079571632049